api/v2: correct ipns api test path
